add update next actions fix car next move nodeId to int

+++ world/src/main/scala/com/sberhack/fun/car/CarNextMove.scala
@@ -1,7 +1,7 @@
 package com.sberhack.fun.car
 
 case class CarNextMove(
-                        nodeId: Long,
+                        nodeId: Int,
                         cashIn: Boolean
                       ) {
